CBP_API/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-storage
python-dotenv